pyproject.toml
[project]
name = "zocketassignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-google-genai>=2.1.6",
    "langchain>=0.3.26",
    "langgraph>=0.5.1",
    "langchain-community>=0.3.27",
    "firecrawl-py>=2.15.0",
    "serpapi>=0.1.5",
    "python-dotenv>=1.1.1",
    "ipykernel>=6.29.5",
    "uvicorn>=0.35.0",
    "fastapi>=0.116.0",
]
